Implement a new type of Constructor that takes a string as argument
The new constructor Matrix(int rows, int cols, string numbers) uses the
integers rows and cols to define the size of the matrix, an take the
data from the string. the data consists of double type numbers,
separated by a blank space. if the number of numbers is too low, it
fills the missing data with 0s